Remember Option
Added a remember checkbox to opening activity. During OnCreate, the app
checks for a local text file, and either grabs a string from it or
creates an empty version of it. The string is split on a '&' character
and the username and password fields are filled with the result.
During login, if the checkbox is checked, the app will grab the values
from the username and password fields and write them into the text file.
Very little error checking or special case handling is done at this
point.