Local Player Information
Added functionality to maintain name and player number of local player.
From entering username on start screen, player's name is held in bundles
and passed down to game activity, where it is held as a member variable.
On calling getCurrentPlayer(), if the local players hasn't been assigned
a role (P1 or P2), the game sets that information locally.
TODO: check current player in game against local player information, and
wait local player if not currently playing