Threaded Select Delay
Added threaded dialog box for bird select menu as well. Currently
doesn't catch the initial load, I'll figure that out soon, but once the
game is under way, the dialog box appears when the other player is
choosing a bird. Once the xml is in place, it will know to skip their
select screen for the local player.
Also, added the threaded call to the cloud class to get the updated
game. Currently, the cloud class loops indefinitely and the app will
hang, but by removing the loop it works fine for returning from the
thread to the game. However, I've discarded my changes to Cloud.java so
as to not conflict with Paul's work